+++ src/random_ray/flat_source_domain.cpp
@@ -249,7 +249,7 @@ void FlatSourceDomain::set_flux_to_flux_plus_source(
     source_regions_.scalar_flux_new(sr, g) /= volume;
     if (settings::run_mode == RunMode::FIXED_SOURCE) {
       source_regions_.scalar_flux_new(sr, g) +=
-        0.5f * source_regions_.external_source(sr, g) *
+        0.5 * source_regions_.external_source(sr, g) *
         source_regions_.volume_sq(sr);
     }
   } else {
@@ -2066,7 +2066,7 @@ void FlatSourceDomain::compute_single_T1(SourceRegionHandle& srh)
     }
 
     // Multiply out sigma_t to correctly compute the derivative term
-    float source_time_derivative =
+    double source_time_derivative =
       A0 * srh.source(g) * sigma_t + srh.source_rhs_bd(g);
 
     double scalar_flux_time_derivative_2 =
@@ -2261,7 +2261,7 @@ void FlatSourceDomain::store_time_step_quantities(bool increment_not_initialize)
           sigma_t =
             sigma_t_[material_offset + g] * source_regions_.density_mult(sr);
         }
-        float source = source_regions_.source_final(sr, g) * sigma_t;
+        double source = source_regions_.source_final(sr, g) * sigma_t;
         add_value_to_bd_vector(source_regions_.source_bd(sr, g), source,
           increment_not_initialize, RandomRay::bd_order_);
       }
